+++ sdk/karma.conf.js
@@ -1,4 +1,4 @@
-const createConfig = require('@iungo/sdk-utils/karma');
+const createConfig = require('@iungopbx/sdk-utils/karma');
 
 module.exports = function(config) {
     config.set(createConfig(config, __dirname));